pyproject.toml
[project]
name = "v2b-syndata"
version = "0.1.0"
description = "Configurable synthetic V2B (Vehicle-to-Building) dataset generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "networkx>=3.2",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "pydantic>=2.5",
]

[project.scripts]
v2b-syndata = "v2b_syndata.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/v2b_syndata"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.6",
    "mypy>=1.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
